no commit message

+++ myclouds-admin/src/main/java/com/gitee/myclouds/admin/modules/module/ModuleService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import com.alibaba.fastjson.JSON;
 import com.gitee.myclouds.admin.domain.mymodule.MyModuleEntity;
 import com.gitee.myclouds.admin.domain.mymodule.MyModuleMapper;
+import com.gitee.myclouds.admin.modules.cache.CacheMiscService;
 import com.gitee.myclouds.common.web.vo.ZTreeNodeVO;
 import com.gitee.myclouds.toolbox.wrap.Dto;
 import com.gitee.myclouds.toolbox.wrap.Dtos;
@@ -27,6 +28,8 @@ public class ModuleService {
 	private MyModuleMapper myModuleMapper;
 	@Autowired
 	private SqlSession sqlSession;
+	@Autowired
+	private CacheMiscService cacheMiscService;
 	
 	/**
 	 * 查询
@@ -66,6 +69,7 @@ public class ModuleService {
 		//拷贝参数对象中的属性到实体对象中
 		MyModuleEntity myModuleEntity = new MyModuleEntity().copyFrom(inDto);
 		myModuleMapper.insert(myModuleEntity);
+		cacheMiscService.cacheModules();
 		outDto = Dtos.newDto().put2("code", "1").put2("msg", "资源模块信息保存成功");
 		return outDto;
 	}
@@ -80,6 +84,7 @@ public class ModuleService {
 		Dto outDto = null;
 		MyModuleEntity myModuleEntity = new MyModuleEntity().copyFrom(inDto);
 		myModuleMapper.updateByKey(myModuleEntity);
+		cacheMiscService.cacheModules();
 		outDto = Dtos.newDto().put2("code", "1").put2("msg", "资源模块信息修改成功");
 		return outDto;
 	}
@@ -92,8 +97,13 @@ public class ModuleService {
 	 */
 	public Dto delete(Dto inDto) {
 		Integer moduleId = inDto.getInteger("id");
+		Integer cnt = sqlSession.selectOne("sql.module.countSubModules", moduleId);
+		if (cnt > 0) {
+			return Dtos.newDto().put2("code", "-1").put2("msg", "该节点包含子节点，不能删除。请先删除子节点。");
+		}
 		myModuleMapper.deleteByKey(moduleId);
 		sqlSession.delete("sql.module.deleteMyRoleModule", moduleId);
+		cacheMiscService.cacheModules();
 		Dto outDto = Dtos.newDto().put2("code", "1").put2("msg", "资源模块删除成功");
 		return outDto;
 	}
